requirements.txt
newspaper4k[all]